Appointment double-click fires event, request #4: http://code.google.com/p/gwt-cal/issues/detail?id=4
Used the addOpenHandler to signify a double click ... since a double-click technically "opens" the event. May change in the future, but chose to re-use existing events and handlers for now.

Also used AppointmentInterface as opposed to Appointment. Not sure I want to do this...

Example:

dayView.addOpenHandler(new OpenHandler<AppointmentInterface>() {
 @Override
 public void onOpen(OpenEvent<AppointmentInterface> event) {

 }
});