got rid of the Appointment.selected variable, in favor of the Calendar.isSelectedAppointment method.
It is possible that an appointment could be displayed on two calendars, where it selected on one calendar and not the other. Also this avoids any possible syncing issues where an appointment thinks it is selected, but the calendar does not. Example: developers sets appointment.selected(true), but does not set calendar.setSelectedAppointment(appointment)