Added the AppointmentStyle class, which is used to store style attributes for different Appointment color combinations. The idea is that color combinations will not be stored in a css file, but will be set in the inline style of the element. This is done for three reasons 1) because it cleans up the style sheets, eliminating 100's of lines of code and 2) fixes the bug where IE isn't rendering multiple css classes correctly in the month view and 3) it makes managing styles WAY easier due to lack of repetitive css.
I then added the AppointmentStyles class. Here you can get a pre-defined style, such as AppointmentStyles.GREEN

The pre-defined AppointmentStyle classes are not actually defined in the AppointmentStyles class. Instead they are defined in the AppointmentTheme class (abstract)... and the actual pre-defined AppointmentStyle classes are created in either the GoogleAppointmentTheme class or the ICalAppointmentTheme class, depending on which theme you implement. AppointmentStyles uses deferred binding to get the appropriate AppointmentTheme and provide those pre-defined styles to the user.