Issue 114 Bug Fix: Do not use a background image if none is specified for the theme/style.
originally posted in http://groups.google.com/group/gwt-cal/browse_thread/thread/e060bbc29d0d6765

I created a new type ThemeAppointmentStyle to abstract the shared behavior of the GoogleAppointmentStyle and ICalAppointmentStyle and use this abstraction in a template method in the DayViewStyleManager and MonthViewStyleManager classes. Two new methods must be overriden by the subclasses, in order to provide an actual implementation for the *ViewStyleManager class to use when styling a single appointment.