Issue 110, Appointment.clone() does not clone start and end dates.
Adding the cloning of the appointment's start and end properties through a new utility method that verifies if the value is not null. Adding a test case for the clone method of the Appointment to verify that some values should be 'clones' (not the same, in the JUnit api sense).